+++ app/routes/components/CopyUiProvider/changelog.codegen.ts
@@ -3,4 +3,5 @@
 
 export const changelog = ''
   + '- 2025-10-24: Initial version.\n'
+  + '- 2025-10-25: Use `useRef` to avoid re-creating zustand store on each render.\n'
   ;