+++ src/components/CopyUiProvider/copy-ui-provider.tsx
@@ -1,4 +1,4 @@
-import { createContext, type FC, useEffect } from "react";
+import { createContext, type FC, useEffect, useRef } from "react";
 import { createStore, type StoreApi } from "zustand";
 import { createJssState, type JssState } from "@/utils/jss";
 import { useJss } from "./use-jss";
@@ -18,12 +18,15 @@ type CopyUiProviderProps = {
 const CopyUiProvider: FC<CopyUiProviderProps> = (props) => {
   const { children } = props;
 
-  const store = createStore<CopyUiStore>()(() => ({
-    jssState: createJssState(),
-  }));
+  const storeRef = useRef<StoreApi<CopyUiStore> | null>(null);
+  if (storeRef.current === null) {
+    storeRef.current = createStore<CopyUiStore>()(() => ({
+      jssState: createJssState(),
+    }));
+  }
 
   return (
-    <CopyUiContext.Provider value={store}>
+    <CopyUiContext.Provider value={storeRef.current}>
       {children}
       <JssSsrCss />
     </CopyUiContext.Provider>