+++ src/components/CopyUiProvider/index.ts
@@ -1,2 +1,4 @@
+// CopyUiProvider from copy-ui @ 2025-10-25.
+
 export { CopyUiProvider } from "./copy-ui-provider";
 export { useJss } from "./use-jss";